Add install target and export cmake config
Add an alias fuser::fuster to match the installed target name. Then the
usage in the downstream project is the same for subdirectory and
installed target.

Create and install a fuserConfig.cmake file for `find_package(fuser)`.
The config target searches for dependencies if the required targets are
not yet present.

Install the header file.